review: move to TestFileIOFactory
