+++ service/common/src/testFixtures/java/org/apache/polaris/service/TestFileIOFactory.java
@@ -16,10 +16,10 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.polaris.service.catalog.io;
+
+package org.apache.polaris.service;
 
 import jakarta.annotation.Nonnull;
-import jakarta.enterprise.context.ApplicationScoped;
 import java.util.Map;
 import java.util.Set;
 import org.apache.iceberg.catalog.TableIdentifier;
@@ -27,35 +27,26 @@ import org.apache.iceberg.io.FileIO;
 import org.apache.polaris.core.context.CallContext;
 import org.apache.polaris.core.persistence.PolarisResolvedPathWrapper;
 import org.apache.polaris.core.storage.PolarisStorageActions;
+import org.apache.polaris.service.catalog.io.FileIOFactory;
 
-/**
- * Interface for providing a way to construct FileIO objects, such as for reading/writing S3.
- *
- * <p>Implementations are available via CDI as {@link ApplicationScoped @ApplicationScoped} beans.
- */
-public interface FileIOFactory {
+/** A FileIOFactory that always returns the same FileIO instance. */
+public class TestFileIOFactory implements FileIOFactory {
 
-  /**
-   * Loads a FileIO implementation for a specific table in the given realm with detailed config.
-   *
-   * <p>This method may obtain subscoped credentials to restrict the FileIO's permissions, ensuring
-   * secure and limited access to the table's data and locations.
-   *
-   * @param callContext the call for which the FileIO is being loaded.
-   * @param ioImplClassName the class name of the FileIO implementation to load.
-   * @param properties configuration properties for the FileIO.
-   * @param identifier the table identifier.
-   * @param tableLocations locations associated with the table.
-   * @param storageActions storage actions allowed for the table.
-   * @param resolvedEntityPath resolved paths for the entities.
-   * @return a configured FileIO instance.
-   */
-  FileIO loadFileIO(
+  private final FileIO fileIO;
+
+  public TestFileIOFactory(@Nonnull FileIO fileIO) {
+    this.fileIO = fileIO;
+  }
+
+  @Override
+  public FileIO loadFileIO(
       @Nonnull CallContext callContext,
       @Nonnull String ioImplClassName,
       @Nonnull Map<String, String> properties,
       @Nonnull TableIdentifier identifier,
       @Nonnull Set<String> tableLocations,
       @Nonnull Set<PolarisStorageActions> storageActions,
-      @Nonnull PolarisResolvedPathWrapper resolvedEntityPath);
+      @Nonnull PolarisResolvedPathWrapper resolvedEntityPath) {
+    return fileIO;
+  }
 }
